+++ src/main/javascript/Security/OauthFacade.js
@@ -1,6 +1,6 @@
 import * as Events from './events';
 import OauthToken from './OauthToken';
-// import OauthConnection from './OauthConnection';
+import OauthConnection from './OauthConnection';
 
 const defaultProtocolVersion = '2.0';
 
@@ -65,6 +65,20 @@ class OauthFacade
   };
 
   /**
+   * Alias for OauthFacade.requestAccess
+   *
+   * @param {string} provider the provider id
+   * @param {{protocolVersion:String}} [options] a map of properties describing the oauth client. Only `protocolVersion` is supported for the moment
+   * @internal
+   * @deprecated
+   * @returns {Promise.<OauthToken, Error>}
+   */
+  async access(provider, options)
+  {
+    return this.requestAccess(provider, options);
+  }
+
+  /**
    * Requests an access token. If `protocolVersion` is not specified in `options` then it defaults to `2.0`
    *
    * @method
@@ -73,7 +87,7 @@ class OauthFacade
    * @param {{protocolVersion:String}} [options] a map of properties describing the oauth client. Only `protocolVersion` is supported for the moment
    * @return {Promise<OauthToken, Error>}
    */
-  async access(provider, options)
+  async requestAccess(provider, options)
   {
     let eventOptions = null;
     if (typeof options === 'object') {